Removed Theme Developer module.
This is removed because no one was using it.

Make sure to uninstall the module BEFORE this commit is merged to the master branch for your site. Since the module file will be removed after the merge, there's no way to uninstall the module after the merge has been done.