Added update instructions that sets the install_language variable.
This variable is used to determine whether or not the language field and filter should be available for the content administration view. The update instructions will make a clever guess to determine which language to use.

If this fails, you should alter the install_language variable manually in the database. It's located in the variables table.