Added Metatag module with default configurations.
The metatag module will add metatags and open graph tags to every page. The default configuration is being used but it's possible to override it.

The module won't be enabled for existing sites. If your site already has the metatags module, you may remove it from sites/all and use the one from the profile instead.