Applied stored patches to Drupal Core.
We've been using these patches prior to Drupal 7.23, but I've forgot to apply them after the upgrade to 7.23.

The biggest difference is the inheritance of install profiles, which we were using when installing Odd Vault.

For sites that has been installed with the Odd Drupal profile, run the following SQL query:

INSERT INTO `variable` (`name`, `value`)
VALUES
	('install_profiles', X'613A313A7B693A303B733A393A226F646464727570616C223B7D');

If you get a "Duplicate entry" failure, the variable already exists and nothing needs to be done. If you don't get any errors, the variable has been added, and you'll need to clear the cache.