Changed Fanta in favor for jReject.
jReject is a module that will warn users with an outdated browser. jReject is much more flexible since it allows for more fine-grained configuration.

Before merging this commit, make sure that the previous module, Fanta, is disabled and uninstalled for every environment. Then run update.php in order to enable and configure jReject.