Changed the symlink to the libraries folder in Odd Vault to an actual folder.
This reverts commit 460d367e7750e523f96063c91497b635fa1ab672.

It's not always possible to deploy symlinks via FTP which means that we can't use that work-around. Since different modules scans the libraries folder in their own way, it's not a good solution to alter the way that the modules work.

We'll need to fallback to the ugly, but safest, solution and keep a duplicated libraries folder for the Odd Vault profile.