Added Masquerade and Masquerade Views.
Added a field to the user administration view that allows users with the appropriate permission to masquerade as another user.
Print a message during init if the user is masquerading, which allows the user to switch back.
Added update instructions that enables the modules for existing sites.

You will need to add the field to the user administration view manually, if the view is overridden for the current site.