Removed the redirect rule in favor for the R4032login module.
The rule which redirected anonymous users to the login page has been removed. This means that the Rules condition site variable module has been removed as well.

The new redirect is handled by the Redirect 403 to User Login module. This will redirect all anonymous users to user/login if the user tries to access a page which generates a 403.

Update instructions:
- Disable, and uninstall the Rules Condition Site Variable module for every environment before merging this commit.
- Run update.php in order to enable and configure the new module.
- Verify that the "Redirect anonymous users to login page" rule has been removed. If it hasn't, remove it manually.