Changed the VBO delete action for users to cancel.
We'll get another step when cancelling acounts that allows us to select the appropriate action. Delete the account and its content, delete the account and make the content belong to the anonymous user, etc.

A regular delete action will simply delete the account and all its content, which isn't desirable.