Disable Field UI for the production environment.
This will add the Field UI module to the Tadaa! configuration in order to enable it for the development and staging environments, but disable it for the production environment.

Run update.php in order to apply the new Tadaa! configuration and to automatically enable/disable the module, depending on the current environment.

Make sure that every feature is in the default state after the update. If a contrib module depends on the Field UI module (unlikely), you will need to manually revert the Tadaa! configuration in order to keep the Field UI module enabled for every environment.