Removed the dependency to the PHP module.
It's often bad practise to use the PHP filter, so we shouldn't encourage people to use it.

Existing sites won't be effected since some sites actually needs this module.

However, you should disable it for existing sites if it's possible. If it is possible, you should also add the Bad Judgement module to the Tadaa! configuration in order to enable it for the environments where Letmein is enabled.