Added a default README.
This file should be used to provide a brief description of the project and the site architecture.

You will need to modify the contents of this file in order to provide actual information! This file is only supposed to be used as a starting point.

Ask around if you can't fill in the blanks by yourself.