Removed the patch for unlimited shortcuts.
This patch is unecessary since it's possible to alter the maximum amount of shortcuts via a variable.

Check the current amount of shortcuts for the site before deploying this commit. You'll need to alter the variable IF there are more than 7 shortcuts.

To do this, simply add a line to settings.site.php, which says "$conf['shortcut_max_slots'] = 10;", or whatever number you prefer.