Added Security Review module.
The module will be enabled for the staging and production environments since the development environment isn't necessary to check and could provide som inaccurate results due to the file system etc.

Run update.php in order to update the Tadaa configuration, and then perform the check for staging and production in order to see if you need to take any actions regarding security.