SED-292: Fixed table reloads not working properly.
The reload URL where incorrect at times when the table was rebuilt during custom AJAX callbacks. This was due to the fact that the URL was different than the offer node path, which made it impossible to reload the page based on the offer. This has been solved by changing the paths. Also fixed logics for reloading the kontaktlista page no matter which row was being edited.